Moved build.py, spatul.py, cross_reference.py to scripts/
